requirements.txt
torch
torchvision
open-clip-torch
pandas
Pillow
ipython
tqdm
ipykernel
nltk
flask
scikit-learn